+++ berserk/types/puzzles.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from typing import Literal, List
-from typing_extensions import TypedDict
+from typing_extensions import NotRequired, TypedDict
 
 from .common import Color
 
@@ -14,6 +14,9 @@ class PuzzleUser(TypedDict):
     name: str
     color: Color
     rating: int
+    flair: NotRequired[str]
+    patron: NotRequired[bool]
+    patronColor: NotRequired[int]
 
 
 class PuzzlePerf(TypedDict):
@@ -42,6 +45,10 @@ class PuzzleInfo(TypedDict):
 class PuzzleData(TypedDict):
     game: PuzzleGame
     puzzle: PuzzleInfo
+
+
+class PuzzleBatchResponse(TypedDict):
+    puzzles: List[PuzzleData]
 
 
 class PuzzleRace(TypedDict):